+++ packages/react/src/components/Form/TextField.tsx
@@ -2,7 +2,7 @@ import { Form } from '@cos-ui/primitives';
 import styled from 'styled-components';
 
 import { TextAreaStyle } from '../TextArea';
-import { TextInputStyle } from '../TextInput';
+import { TextInputStyle } from '../TextInput/TextInput';
 import { FieldStyle } from './Field';
 
 const TextField = styled(Form.TextField)`
