+++ packages/react/src/components/TextInput/index.ts
@@ -1,0 +1,1 @@
+export { default as TextInput } from './TextInput';
